@@ -1,26 +1,33 @@
-[tool.poetry]
+[build-system]
+requires = ["setuptools >= 77.0.3"]
+build-backend = "setuptools.build_meta"
+
+[tool.setuptools.packages.find]
+include = ["hairpin2*"]
+
+[project]
 name = "hairpin2"
-version = "1.2.0"
-description = "CLI implementation of the artefact detection algorithm concieved by Ellis et al, 2020."
-authors = ["Alex Byrne <ab63@sanger.ac.uk>"]
+version = "2.0.0"
+description = "CLI for running read-aware artefactual variant flagging algorithms over VCFs"
+authors = [
+  {name = "Alex Byrne"}
+]
 license = "MIT"
 readme = "README.md"
-
-[tool.poetry.dependencies]
-python = "^3.10"
-pysam = "^0.22"
-
-[tool.poetry.plugins."console_scripts"]
-hairpin2 = "hairpin2.main:main_cli"
-
-[tool.poetry.group.dev.dependencies]
-pytest = "^8.2.2"
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
-
-[tool.pytest.ini_options]
-markers = [
-  "validate: scientific validation tests"
+dependencies = [
+  "pysam>=0.22, ==0.*",
+  "pydantic>=2.11.7, ==2.*",
+  "click>=8.2.1, ==8.*",
+  "click-option-group>=0.5.7, ==0.*"
 ]
+requires-python = ">= 3.12"
+
+[project.optional-dependencies]
+dev = [
+  "pytest==8.*",
+  "pytest-cov==6.*"
+]
+
+[project.scripts]
+hairpin2 = "hairpin2.main:hairpin2"
+
