@@ -1,6 +1,6 @@
 # hairpin2
 #
-# Copyright (C) 2024 Genome Research Ltd.
+# Copyright (C) 2024, 2025 Genome Research Ltd.
 #
 # Author: Alex Byrne <ab63@sanger.ac.uk>
 #
@@ -21,45 +21,21 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
-
 from enum import IntEnum, Flag
-import logging
-import sys
-from hairpin2 import constants as c
-
-
-def cleanup(code: int = c.EXIT_FAILURE, msg: None | str = None) -> None:
-    if code != c.EXIT_SUCCESS and msg:
-        logging.error(msg)
-    if code == c.EXIT_SUCCESS:
-        logging.info('hairpin complete')
-    sys.exit(code)
-
-
-def has_duplicates(
-    l: list
-) -> bool:
-    return len(l) != len(set(l))
-
-
-def lists_not_equal(
-    l1: list | set,
-    l2: list | set
-) -> bool:
-    return sorted(l1) != sorted(l2)
+# pyright: reportExplicitAny=false
 
 
 def print_flag(
     print_enum: Flag,
 ) -> None:
-    pl = []
+    pl: list[str] = []
     for e in print_enum:
         vs = '-'.join([str(int(e.value)), str(hex(e.value)), str(bin(e.value))])
-        pl.append(': '.join([str(e), vs]))
+        pl.append(f'{e}: {vs}')
     print(pl)
 
 
 def print_enum(
     print_enum: IntEnum
 ) -> None:
-    print([e for e in print_enum])  # type: ignore - iterating works fine
+    print([e for e in print_enum])  # pyright: ignore[reportGeneralTypeIssues] - iterating works fine
