@@ -1,6 +1,6 @@
 # hairpin2
 #
-# Copyright (C) 2024 Genome Research Ltd.
+# Copyright (C) 2024, 2025 Genome Research Ltd.
 #
 # Author: Alex Byrne <ab63@sanger.ac.uk>
 #
@@ -21,18 +21,13 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
-
-from hairpin2.main import flag_read_alt
-from hairpin2 import constants as c
+from hairpin2.readqc import qc_read_alt_specific, ValidatorFlags
 import pysam
 from pysam.libcalignedsegment import SAM_FLAGS as s
 import copy
-import pytest
+from pytest import raises
 
 
-# BASIS PATH TESTING (ish)
-# test every node and edge at least once
-# ----
 # perfect read pair:
 r = pysam.AlignedSegment()
 r.query_name = 'read1'
@@ -47,21 +42,19 @@ r.cigarstring = '10M'
 r.set_tag('MC', '10M')
 
 
-@pytest.mark.validate
 def test_path_unsupported_mut_type():
-    with pytest.raises(ValueError):
-        flag_read_alt(read=r,
+    with raises(ValueError):
+        _ = qc_read_alt_specific(read=r,
                       vcf_start=99,
                       vcf_stop=100,
                       alt='A',
-                      mut_type='8',  # type: ignore
+                      mut_type='8',  # pyright: ignore[reportArgumentType]
                       min_basequal=25)
 
 
-@pytest.mark.validate
 def test_path_sub_not_aligned():
-    expected = c.ValidatorFlags.NOT_ALIGNED.value
-    result = flag_read_alt(read=r,
+    expected = ValidatorFlags.NOT_ALIGNED
+    result = qc_read_alt_specific(read=r,
                            vcf_start=200,
                            vcf_stop=100,
                            alt='A',
@@ -70,11 +63,10 @@ def test_path_sub_not_aligned():
     assert expected == result
 
 
-@pytest.mark.validate
 def test_path_bad_sub():
-    expected = (c.ValidatorFlags.NOT_ALT.value
-                | c.ValidatorFlags.BASEQUAL.value)
-    result = flag_read_alt(read=r,
+    expected = (ValidatorFlags.NOT_ALT
+                | ValidatorFlags.BASEQUAL)
+    result = qc_read_alt_specific(read=r,
                            vcf_start=99,
                            vcf_stop=100,
                            alt='T',
@@ -83,10 +75,9 @@ def test_path_bad_sub():
     assert expected == result
 
 
-@pytest.mark.validate
 def test_path_good_sub():
-    expected = c.ValidatorFlags.CLEAR.value
-    result = flag_read_alt(read=r,
+    expected = ValidatorFlags.CLEAR
+    result = qc_read_alt_specific(read=r,
                            vcf_start=99,
                            vcf_stop=100,
                            alt='A',
@@ -95,10 +86,9 @@ def test_path_good_sub():
     assert expected == result
 
 
-@pytest.mark.validate
 def test_path_del_short():
-    expected = c.ValidatorFlags.SHORT.value
-    result = flag_read_alt(read=r,
+    expected = ValidatorFlags.SHORT
+    result = qc_read_alt_specific(read=r,
                            vcf_start=99,
                            vcf_stop=110,
                            alt='.',
@@ -107,10 +97,9 @@ def test_path_del_short():
     assert expected & result
 
 
-@pytest.mark.validate
 def test_path_del_bad_op():
-    expected = c.ValidatorFlags.BAD_OP.value
-    result = flag_read_alt(read=r,
+    expected = ValidatorFlags.BAD_OP
+    result = qc_read_alt_specific(read=r,
                            vcf_start=99,
                            vcf_stop=101,
                            alt='C',
@@ -120,12 +109,11 @@ def test_path_del_bad_op():
 
 
 # 2bp del
-@pytest.mark.validate
 def test_path_good_del():
-    expected = c.ValidatorFlags.CLEAR.value
+    expected = ValidatorFlags.CLEAR
     rc = copy.deepcopy(r)
     rc.cigarstring = '4M2D6M'
-    result = flag_read_alt(read=rc,
+    result = qc_read_alt_specific(read=rc,
                            vcf_start=98,
                            vcf_stop=101,
                            alt='CC',
@@ -134,10 +122,9 @@ def test_path_good_del():
     assert expected == result
 
 
-@pytest.mark.validate
 def test_path_ins_not_aligned():
-    expected = c.ValidatorFlags.NOT_ALIGNED.value
-    result = flag_read_alt(read=r,
+    expected = ValidatorFlags.NOT_ALIGNED
+    result = qc_read_alt_specific(read=r,
                            vcf_start=200,
                            vcf_stop=100,
                            alt='A',
@@ -146,10 +133,9 @@ def test_path_ins_not_aligned():
     assert expected == result
 
 
-@pytest.mark.validate
 def test_path_ins_short():
-    expected = c.ValidatorFlags.SHORT.value
-    result = flag_read_alt(read=r,
+    expected = ValidatorFlags.SHORT
+    result = qc_read_alt_specific(read=r,
                            vcf_start=99,
                            vcf_stop=100,
                            alt='ATTTTTTTTTTTTTT',
@@ -158,10 +144,9 @@ def test_path_ins_short():
     assert expected == result
 
 
-@pytest.mark.validate
 def test_path_bad_ins():
-    expected = (c.ValidatorFlags.BAD_OP.value | c.ValidatorFlags.NOT_ALT.value)
-    result = flag_read_alt(read=r,
+    expected = (ValidatorFlags.BAD_OP | ValidatorFlags.NOT_ALT)
+    result = qc_read_alt_specific(read=r,
                            vcf_start=99,
                            vcf_stop=100,
                            alt='AC',
@@ -170,12 +155,11 @@ def test_path_bad_ins():
     assert expected == result
 
 
-@pytest.mark.validate
 def test_path_good_ins():
-    expected = c.ValidatorFlags.CLEAR.value
+    expected = ValidatorFlags.CLEAR
     rc = copy.deepcopy(r)
     rc.cigarstring = '5M2I3M'
-    result = flag_read_alt(read=rc,
+    result = qc_read_alt_specific(read=rc,
                            vcf_start=99,
                            vcf_stop=100,
                            alt='AA',
