@@ -1,6 +1,6 @@
 # hairpin2
 #
-# Copyright (C) 2024 Genome Research Ltd.
+# Copyright (C) 2024, 2025 Genome Research Ltd.
 #
 # Author: Alex Byrne <ab63@sanger.ac.uk>
 #
@@ -21,18 +21,12 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
-
-from hairpin2.main import flag_read_broad
-from hairpin2 import constants as c
+from hairpin2.readqc import qc_read_broad, ValidatorFlags
 import pysam
 from pysam.libcalignedsegment import SAM_FLAGS as s
 import copy
-import pytest
 
 
-# BASIS PATH TESTING (ish)
-# test every node and edge at least once
-# ----
 # perfect read pair:
 r = pysam.AlignedSegment()
 r.query_name = 'read1'
@@ -47,74 +41,68 @@ r.cigarstring = '10M'
 r.set_tag('MC', '10M')
 
 
-@pytest.mark.validate
 def test_path_clear():
-    expected = c.ValidatorFlags.CLEAR.value
-    result = flag_read_broad(read=r,
+    expected = ValidatorFlags.CLEAR
+    result = qc_read_broad(read=r,
                              vcf_start=99,
                              min_mapqual=11,
                              min_clipqual=35)
     assert expected == result
 
 
-@pytest.mark.validate
 def test_path_missing_mc():
-    expected = c.ValidatorFlags.READ_FIELDS_MISSING.value
+    expected = ValidatorFlags.READ_FIELDS_MISSING
     rc = copy.deepcopy(r)
     rc.set_tag('MC', None)
-    result = flag_read_broad(read=rc,
+    result = qc_read_broad(read=rc,
                              vcf_start=99,
                              min_mapqual=11,
                              min_clipqual=35)
     assert expected == result
 
 
-@pytest.mark.validate
 def test_path_missing_field():
-    expected = c.ValidatorFlags.READ_FIELDS_MISSING.value
+    expected = ValidatorFlags.READ_FIELDS_MISSING
     rc = copy.deepcopy(r)
     rc.cigarstring = None
-    result = flag_read_broad(read=rc,
+    result = qc_read_broad(read=rc,
                              vcf_start=99,
                              min_mapqual=11,
                              min_clipqual=35)
     assert expected == result
 
 
-@pytest.mark.validate
 def test_path_set_flag_mapqual_clipqual():
-    expected = (c.ValidatorFlags.FLAG.value
-                | c.ValidatorFlags.MAPQUAL.value
-                | c.ValidatorFlags.CLIPQUAL.value)
+    expected = (ValidatorFlags.FLAG
+                | ValidatorFlags.MAPQUAL
+                | ValidatorFlags.CLIPQUAL)
     rc = copy.deepcopy(r)
     rc.flag = s.FQCFAIL  # 0x200
     rc.cigarstring = '1S9M'
-    result = flag_read_broad(read=rc,
+    result = qc_read_broad(read=rc,
                              vcf_start=99,
                              min_mapqual=30,
                              min_clipqual=40)
     assert expected == result
 
 
-@pytest.mark.validate
 def test_path_overlap():
-    expected = c.ValidatorFlags.OVERLAP.value
+    expected = ValidatorFlags.OVERLAP
     rc = copy.deepcopy(r)
     rc.flag = s.FPAIRED | s.FPROPER_PAIR | s.FREAD2  # 0x80
-    result = flag_read_broad(read=rc,
+    result = qc_read_broad(read=rc,
                              vcf_start=99,
                              min_mapqual=11,
                              min_clipqual=40)
     assert expected == result
 
 
-@pytest.mark.validate
 def test_path_no_overlap():
-    expected = c.ValidatorFlags.CLEAR.value
+    expected = ValidatorFlags.CLEAR
     rc = copy.deepcopy(r)
     rc.flag = s.FPAIRED | s.FPROPER_PAIR | s.FREAD2  # 0x80
     rc.set_tag('MC', '3M')
-    result = flag_read_broad(read=rc,
+    result = qc_read_broad(read=rc,
                              vcf_start=99,
                              min_mapqual=11,
                              min_clipqual=40)
