@@ -1,6 +1,6 @@
 # hairpin2
 #
-# Copyright (C) 2024 Genome Research Ltd.
+# Copyright (C) 2024, 2025 Genome Research Ltd.
 #
 # Author: Alex Byrne <ab63@sanger.ac.uk>
 #
@@ -21,15 +21,11 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
-
 from hairpin2.ref2seq import ref2querypos
-import pytest
+from pytest import raises
 import pysam
 
 
-# BASIS PATH TESTING (ish)
-# test every node and edge at least once
-# ----
 # perfect read pair:
 r = pysam.AlignedSegment()
 r.query_name = 'read1'
@@ -44,13 +40,11 @@ r.cigarstring = '10M'
 r.set_tag('MC', '10M')
 
 
-@pytest.mark.validate
-def test_path_indexerror():
-    with pytest.raises(IndexError):
-        ref2querypos(r, 1000)
+def test_path_error():
+    with raises(ValueError):
+        _ = ref2querypos(r, 1000)
 
 
-@pytest.mark.validate
 def test_path_good():
     expected = 5
     result = ref2querypos(r, 100)
