@@ -1,6 +1,6 @@
 # hairpin2
 #
-# Copyright (C) 2024 Genome Research Ltd.
+# Copyright (C) 2024, 2025 Genome Research Ltd.
 #
 # Author: Alex Byrne <ab63@sanger.ac.uk>
 #
@@ -23,19 +23,14 @@
 # SOFTWARE.
 
 from hairpin2.ref2seq import ref_end_via_cigar
-import pytest
+from pytest import raises
 
 
-# BASIS PATH TESTING (ish)
-# test every node and edge at least once
-# ----
-@pytest.mark.validate
 def test_path_valueerror():
-    with pytest.raises(ValueError):
-        ref_end_via_cigar('30M2I* ,', 0)
+    with raises(ValueError):
+        _ = ref_end_via_cigar('30M2I* ,', 0)
 
 
-@pytest.mark.validate
 def test_path_normal():
     expected = 50
     result = ref_end_via_cigar('20M10I30M', 0)
